Update UI components with required fields and fix on Select options font family
