+++ src/app/(private)/customers/[id]/PanelSegmentation.tsx
@@ -55,7 +55,7 @@ export default function PanelSegmentation({
             >
               <Select ml={{ md: '0', lg: '0', xl: '4rem' }} placeholder="Definir o Segmento" {...register('segment')}>
                 {segments?.segments.map((segment) => (
-                  <option key={`${segment.name}-${segment.id}`} value={segment.id}>
+                  <option key={`${segment.name}-${segment.id}`} value={segment.id} style={{ fontFamily: 'sans-serif' }}>
                     {segment.name}
                   </option>
                 ))}
@@ -77,7 +77,7 @@ export default function PanelSegmentation({
             >
               <Select ml={{ md: '0', lg: '0', xl: '4rem' }} placeholder="Definir o Parceiro" {...register('partner')}>
                 {partners?.partners.map((partner) => (
-                  <option key={`${partner.name}-${partner.id}`} value={partner.id}>
+                  <option key={`${partner.name}-${partner.id}`} value={partner.id} style={{ fontFamily: 'sans-serif' }}>
                     {partner.name}
                   </option>
                 ))}
@@ -103,7 +103,7 @@ export default function PanelSegmentation({
                 {...register('flag')}
               >
                 {flags?.flags.map((flag) => (
-                  <option key={`${flag.name}-${flag.id}`} value={flag.id}>
+                  <option key={`${flag.name}-${flag.id}`} value={flag.id} style={{ fontFamily: 'sans-serif' }}>
                     {flag.name}
                   </option>
                 ))}
@@ -131,7 +131,7 @@ export default function PanelSegmentation({
                 {...register('brand')}
               >
                 {brands?.brands.map((brand) => (
-                  <option key={`${brand.name}-${brand.id}`} value={brand.id}>
+                  <option key={`${brand.name}-${brand.id}`} value={brand.id} style={{ fontFamily: 'sans-serif' }}>
                     {brand.name}
                   </option>
                 ))}
