+++ src/theme/breakpoints.ts
@@ -1,0 +1,13 @@
+const breakpoints = {
+    base: '0px',
+    sm: '320px',
+    md: '768px',
+    lg: '960px',
+    xl: '1200px',
+    '2xl': '1536px',
+    '3xl': '1920px',
+    '4xl': '2560px',
+    '5xl': '3840px',
+};
+
+export { breakpoints };
